requirements.txt
pyyaml>=6.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0
httpx>=0.27.0
pytest>=8.0
